# api/public.openapi.yaml
openapi: 3.0.3
info:
    title: WEBYCP Public API
    version: 0.7.0
    description: Public REST API used by the WEBYCP web application.
paths:
...
    /api/v1/databases/{databaseId}:
        delete:
            operationId: deleteDatabase
            summary: Delete a MySQL database
            security: [{ cookieAuth: [] }]
            parameters:
                - $ref: "#/components/parameters/DatabaseID"
                - $ref: "#/components/parameters/CSRFToken"
            responses:
                "202":
                    description: Database deletion queued.
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/DatabaseJobResponse" }
                "401": { $ref: "#/components/responses/UnauthorizedError" }
                "403": { $ref: "#/components/responses/ForbiddenError" }
                "404": { $ref: "#/components/responses/NotFoundError" }
                "409": { $ref: "#/components/responses/ConflictError" }
                "500": { $ref: "#/components/responses/InternalError" }
...
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: webycp_session
    parameters:
...
        CSRFToken:
            name: X-CSRF-Token
            in: header
            required: true
            schema:
                type: string
...
        DatabaseID:
            name: databaseId
            in: path
            required: true
            schema: { type: string }
...
    responses:
...
        UnauthorizedError:
            description: Authentication is required.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        ForbiddenError:
            description: The request is not allowed.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        NotFoundError:
            description: The resource does not exist.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        ConflictError:
            description: The request conflicts with current state.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
...
        InternalError:
            description: An internal error occurred.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
    schemas:
...
        Job:
            type: object
            additionalProperties: false
            required: [id, kind, status, payload, attempts, maxAttempts, createdAt]
            properties:
                id:
                    type: string
                nodeId:
                    type: string
                    nullable: true
                userId:
                    type: string
                    nullable: true
                kind:
                    type: string
                status:
                    type: string
                    enum: [queued, running, succeeded, failed]
                payload:
                    type: object
                    additionalProperties: true
                attempts:
                    type: integer
                    format: int64
                maxAttempts:
                    type: integer
                    format: int64
                error:
                    type: string
                    nullable: true
                createdAt:
                    type: string
                    format: date-time
                startedAt:
                    type: string
                    format: date-time
                    nullable: true
                finishedAt:
                    type: string
                    format: date-time
                    nullable: true
...
        Database:
            type: object
            additionalProperties: false
            required: [id, accountId, nodeId, name, systemName, status, createdAt, updatedAt]
            properties:
                id: { type: string }
                accountId: { type: string }
                nodeId: { type: string }
                name: { type: string }
                systemName: { type: string }
                status: { type: string, enum: [pending, active, error] }
                createdAt: { type: string, format: date-time }
                updatedAt: { type: string, format: date-time }
...
        DatabaseJobResponse:
            type: object
            additionalProperties: false
            required: [database, job]
            properties:
                database: { $ref: "#/components/schemas/Database" }
                job: { $ref: "#/components/schemas/Job" }
...
        ErrorResponse:
            type: object
            additionalProperties: false
            required: [code, message]
            properties:
                code:
                    type: string
                message:
                    type: string
                fields:
                    type: object
                    additionalProperties:
                        type: string